cambios insert

+++ src/main/java/com/example/activate/service/CalcularKcalService.java
@@ -4,5 +4,5 @@ import com.example.activate.service.CalcularKcalServiceImpl.Sexo;
 import com.example.activate.service.CalcularKcalServiceImpl.NivelActividad;
 
 public interface CalcularKcalService {
-    double calcularCalorias(Sexo sexo, int edad, double peso, double altura, NivelActividad nivelActividad);
+    float calcularCalorias(Sexo sexo, int edad, double peso, double altura, NivelActividad nivelActividad);
 }
